WT-609: Removed all remnant of API/1
