+++ client/Dockerfile
@@ -12,7 +12,6 @@ COPY --from=install /usr/app/package*.json ./
 COPY . .
 
 ARG ENV
-ARG REACT_APP_API_1_URL
 ARG REACT_APP_API_2_URL
 ARG REACT_APP_AUTH0_AUDIENCE
 ARG REACT_APP_AUTH0_CLIENT_ID
@@ -20,7 +19,6 @@ ARG REACT_APP_AUTH0_URL
 ARG REACT_APP_PUBLIC_STRIPE_KEY
 ARG REACT_APP_ROOT_URL
 
-ENV REACT_APP_API_1_URL=${REACT_APP_API_1_URL}
 ENV REACT_APP_API_2_URL=${REACT_APP_API_2_URL}
 ENV REACT_APP_AUTH0_AUDIENCE=${REACT_APP_AUTH0_AUDIENCE}
 ENV REACT_APP_AUTH0_CLIENT_ID=${REACT_APP_AUTH0_CLIENT_ID}
