+++ docker-compose.yml
@@ -38,7 +38,6 @@ services:
       REACT_APP_AUTH0_URL: ${AUTH0_URL}
       REACT_APP_AUTH0_CLIENT_ID: ${AUTH0_CLIENT_ID}
       REACT_APP_AUTH0_AUDIENCE: ${AUTH0_AUDIENCE}
-      REACT_APP_API_1_URL: ${ROOT_URL}/api/1
       REACT_APP_API_2_URL: ${ROOT_URL}/api/2
       ENV: local
     volumes:
