+++ gcp/scripts/build-client.sh
@@ -24,7 +24,6 @@ URL=${ROOT_URL:-$SERVER_REVISION}
   --destination=${ARTIFACTS}/client-img-${ENV}:${SHORT_SHA} \
   --cache=true \
   --cache-repo=${KANIKO_CACHE} \
-  --build-arg REACT_APP_API_1_URL=${URL}/api/1 \
   --build-arg REACT_APP_API_2_URL=${URL}/api/2 \
   --build-arg REACT_APP_AUTH0_AUDIENCE=${URL} \
   --build-arg REACT_APP_AUTH0_CLIENT_ID=${AUTH0_CLIENT_ID} \
