+++ server/.eslintrc.json
@@ -1,7 +1,4 @@
 {
-  "ignorePatterns": [
-    "src/prisma/**/*.ts"
-  ],
   "env": {
     "browser": true,
     "es2021": true
