--- server/src/api/saved_reports/report.ts
@@ -1,5 +1,0 @@
-// this defines the type of data that a saved report expects
-export interface Report {
-  table_name: string;
-  query_attr: string;
-}
