+++ server/src/controllers/eventInstanceController.service.ts
@@ -151,6 +151,24 @@ export const validateDateAndTime = (date: string, time: string) => {
   };
 };
 
+// converts the databases integer representation of a date into a JS date object
+// 20220512 -> 2022-05-12 == May 12 2022
+export const parseIntToDate = (d : number) => {
+  const year = d / 10000 | 0;
+  d -= year *10000;
+  const month = d / 100 | 0;
+  const day = d - month*100;
+  // month-1 because the month field is 0-indexed in JS
+  return new Date(year, month-1, day);
+};
+
+export const getDate = (time: string, date: number) => {
+  const [hour, min] = time.split('T')[1].split(':');
+  const newDate = parseIntToDate(date);
+  newDate.setHours(+hour, +min);
+  return newDate.toJSON();
+};
+
 export const updateShowing = async (
     prisma: ExtendedPrismaClient,
     oldEvent: LoadedEventInstance,
